+++ IcVibracoes.Core/Mapper/IMappingResolver.cs
@@ -1,5 +1,6 @@
 ﻿using IcVibracoes.Common.Classes;
 using IcVibracoes.Core.Models.BeamCharacteristics;
+using IcVibracoes.DataContracts.FiniteElement;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -11,7 +12,7 @@ namespace IcVibracoes.Core.Mapper
     public interface IMappingResolver
     {
         /// <summary>
-        /// It's responsible to build the force vector.
+        /// Thid method builds the force vector.
         /// </summary>
         /// <param name="forces"></param>
         /// <param name="degreesFreedomMaximum"></param>
@@ -19,7 +20,7 @@ namespace IcVibracoes.Core.Mapper
         Task<double[]> BuildForceVector(List<Force> forces, uint degreesFreedomMaximum);
 
         /// <summary>
-        /// It's responsible to build the electrical charge vector.
+        /// Thid method builds the electrical charge vector.
         /// </summary>
         /// <param name="electricalCharges"></param>
         /// <param name="degreesFreedomMaximum"></param>
@@ -30,7 +31,8 @@ namespace IcVibracoes.Core.Mapper
         /// Thid method builds the fastenings of the beam.
         /// </summary>
         /// <param name="fastenings"></param>
+        /// <param name="response"></param>
         /// <returns></returns>
-        Task<IDictionary<uint, FasteningType>> BuildFastenings(List<Fastening> fastenings);
+        Task<IDictionary<uint, FasteningType>> BuildFastenings(List<Fastening> fastenings, FiniteElementResponse response);
     }
 }